third_party/blink/web_tests/virtual/select-parsing-filtering-disabled/fast/forms/select/customizable-select/button-or-datalist-in-select-usecounter-expected.txt

CONSOLE WARNING: An <input> tag was parsed within an open <select> tag which caused a </select> end tag to automatically be inserted. Please add the missing </select> end tag, since this behavior may change in future versions of this browser, possibly causing breakage on this page.
This is a testharness.js-based test.
[FAIL] UseCounter is counted when putting button in select.
  assert_true: button in select expected true got false
[FAIL] UseCounter is counted when putting datalist in select.
  assert_true: datalist in select expected true got false
[FAIL] UseCounter for <div> in <select>.
  assert_true: select parser dropped tag expected true got false
[FAIL] UseCounter is counted when putting input in select.
  assert_true: expected true got false
Harness: the test ran to completion.
